refactor: data model -> data entity로 패키지명 변경
