+++ android/app/src/main/java/com/woowacourse/ody/data/entity/join/JoinRequest.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.data.model.join
+package com.woowacourse.ody.data.entity.join
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
