+++ android/app/src/main/java/com/woowacourse/ody/data/remote/repository/DefaultJoinRepository.kt
@@ -1,7 +1,7 @@
 package com.woowacourse.ody.data.remote.repository
 
-import com.woowacourse.ody.data.model.join.JoinRequest
-import com.woowacourse.ody.data.model.meeting.MeetingResponse
+import com.woowacourse.ody.data.entity.join.JoinRequest
+import com.woowacourse.ody.data.entity.meeting.MeetingResponse
 import com.woowacourse.ody.data.remote.RetrofitClient
 import com.woowacourse.ody.data.remote.service.JoinService
 import com.woowacourse.ody.domain.repository.JoinRepository
