+++ android/app/src/main/java/com/woowacourse/ody/presentation/completion/JoinCompleteActivity.kt
@@ -7,8 +7,8 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.woowacourse.ody.OdyApplication
 import com.woowacourse.ody.R
-import com.woowacourse.ody.data.model.join.JoinRequest
-import com.woowacourse.ody.data.model.meeting.MeetingRequest
+import com.woowacourse.ody.data.entity.join.JoinRequest
+import com.woowacourse.ody.data.entity.meeting.MeetingRequest
 import com.woowacourse.ody.data.remote.repository.DefaultJoinRepository
 import com.woowacourse.ody.data.remote.repository.DefaultMeetingRepository
 import com.woowacourse.ody.presentation.notificationlog.NotificationLogActivity
